Refactored CirclesDiagramView

+++ circlesdiagramview/src/main/java/com/bankcalendar/circlesview/util/Converter.kt
@@ -11,12 +11,12 @@ fun spToPx(sp: Float, context: Context): Float {
     return sp * context.resources.displayMetrics.scaledDensity
 }
 
-fun dpToPx(dp: Float, context: Context): Int {
+fun dpToPx(dp: Float, context: Context): Float {
     return TypedValue.applyDimension(
         TypedValue.COMPLEX_UNIT_DIP,
         dp,
         context.resources.displayMetrics
-    ).toInt()
+    )
 }
 
 fun pxToSp(px: Float, context: Context): Float {
